+++ Shopx/settings.py
@@ -1,14 +1,4 @@
-"""
-Django settings for Shopx project.
-
-Generated by 'django-admin startproject' using Django 4.2.10.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.2/topics/settings/
-
-For the full list of settings and their values, see
-https://docs.djangoproject.com/en/4.2/ref/settings/
-"""
+from celery.schedules import crontab
 import os
 from pathlib import Path
 
@@ -182,3 +172,5 @@ REST_FRAMEWORK = {
     'DEFAULT_SCHEMA_CLASS': 'drf_spectacular.openapi.AutoSchema'
 
 }
+
+
