+++ user_profiles/models.py
@@ -4,7 +4,6 @@ from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
 
 
 from product.models import Product
-from .validators import validator_sms_code
 from .usermanager import CustomUserManager
 
 class CustomUser(AbstractBaseUser, PermissionsMixin):
@@ -12,7 +11,7 @@ class CustomUser(AbstractBaseUser, PermissionsMixin):
     username = models.CharField(max_length= 30, verbose_name="Имя",null=True, blank=True)
     surname = models.CharField(max_length= 30, verbose_name="Фамилия",null=True, blank=True)
     email_or_phone = models.CharField(max_length= 30,unique = True,null= True, blank=True)
-    code = models.CharField(max_length=6, validators = [validator_sms_code])
+    code = models.CharField(max_length=6)
     created_at = models.DateField(auto_now=True)
     is_active = models.BooleanField(default=False)
     is_staff = models.BooleanField(default=False)
