--- user_profiles/permissions.py
@@ -1,14 +1,0 @@
-from rest_framework.permissions import BasePermission
-from .models import *
-
-class IsSeller(BasePermission):
-    def has_permission(self, request, view):
-        return request.user.is_authenticated and request.user.is_seller
-
-class IsBuyer(BasePermission):
-    def has_permission(self, request, view):
-        return request.user.is_authenticated
-
-# class IsWholeseller(BasePermission):
-#     def has_permission(self, request, view):
-#         return request.user.is_authenticated and request.user.is_wholeseller
