+++ user_profiles/urls.py
@@ -1,4 +1,3 @@
-from django.contrib import admin
 from django.urls import path
 from .views import *
 from rest_framework_simplejwt.views import TokenObtainPairView,TokenRefreshView
@@ -8,7 +7,6 @@ urlpatterns = [
     path('login/', TokenObtainPairView.as_view()),
 
     path('seller/register',SellerRegisterView.as_view(), name='seller-register'),
-    # path('wholeseller/register',WholeSellerRegisterView.as_view(), name='wholeseller-register'),
 
     path('refresh-token/', TokenRefreshView.as_view()),
     path('send-code-to-email/', ForgetPasswordSendCodeView.as_view(), name='send_password_reset_code'),
