+++ user_profiles/usermanager.py
@@ -1,6 +1,4 @@
-import unicodedata
 from django.contrib.auth.models import BaseUserManager
-from django.contrib.auth.hashers import make_password
 
 class CustomUserManager(BaseUserManager):
